Add 'role' field to Users
This commit adds the field 'role' to the admin page of Users to create and edit Users with the new field. It also includes additions on the User model to validate and on UserDecorator to 'humanize' the field. Some changes on the User spec and factory were made to test the new field.